ml_image_moderator/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
opennsfw2==0.14.0
tensorflow==2.19.0
boto3==1.37.33
botocore==1.37.33

